cms_accescom/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
graphene==1.4.2
graphene-django==1.3
graphql-core==1.1
graphql-relay==0.5.0
html5lib==1.1
idna==2.10
iso8601==0.1.13
jdcal==1.4.1
l18n==2018.5
openpyxl==3.0.5
Pillow==7.2.0
pkg-resources==0.0.0
promise==2.0.2
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
typing==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.10.2
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.6
xlwt==1.3.0
